requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
Django==3.1.7
djangorestframework==3.12.2
drf-yasg==1.20.0
glue==0.13
html5lib==1.0.1
idna==2.10
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.9.6
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.9
pep517==0.8.2
Pillow==8.1.1
progress==1.5
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-decouple==3.4
pytoml==0.1.21
pytz==2021.1
requests==2.25.1
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.14.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
webencodings==0.5.1
